refactor(console): pin MOD to the catalog, finish the id cleanup
MOD was hand-written next to 21 independently declared module ids, so a
typo on either side still compiled and silently missed the module blob at
runtime -- the exact failure the map exists to prevent. It stays a literal
map (ModuleDef.id is typed string, so deriving the values would erase the
string-literal types the stores key off) and two tests bind it to the
catalog in both directions instead.

Also: the last three `def.id === 'triggers'` literals on the generic module
page now read MOD.triggers; nav.test.ts passed '/' where the parameter is
SectionId, which only slipped through because bun test does not typecheck
shared; the settings page's GRANTABLE_SECTIONS import sat below code; and
catalog/tokens.ts is now catalog/rehearsal-tokens.ts -- "tokens" already
means auth tokens and styles/tokens.css in this repo. The obvious
preview-samples.ts name is not available: assert-demo-gated.mjs treats any
path part matching demo/fixture/sample/seed/mock as a fixture module and
fails the build on a static import of one.
