+++ console/shared/lib/catalog/rehearsal-tokens.ts
