+++ console/shared/lib/module-catalog.test.ts
@@ -2,12 +2,31 @@
 // Proprietary. No license granted. See LICENSE.md.
 
 import { describe, expect, test } from 'bun:test';
-import { MODULE_CATALOG, moduleDef, moduleDelegateSections } from './types';
+import { MOD, MODULE_CATALOG, moduleDef, moduleDelegateSections } from './types';
 
 describe('module catalog', () => {
   test('has unique ids', () => {
     const ids = MODULE_CATALOG.map((def) => def.id);
     expect(new Set(ids).size).toBe(ids.length);
+  });
+
+  // MOD is a hand-written id map sitting next to 21 independently-declared
+  // module ids (each catalog/<name>.ts owns its own `id: '<name>'`). A typo
+  // in either place used to compile fine and just silently miss the module
+  // blob at runtime. These two directions catch a renamed id on one side and
+  // a new module that never got a MOD key on the other.
+  test('every MOD entry names a real catalog module', () => {
+    for (const [key, value] of Object.entries(MOD)) {
+      expect(key).toBe(value);
+      expect(moduleDef(value)).toBeDefined();
+    }
+  });
+
+  test('every catalog module has a MOD key', () => {
+    const catalogIds = MODULE_CATALOG.map((def) => def.id).sort();
+    const modValues = new Set(Object.values(MOD));
+    const missing = catalogIds.filter((id) => !modValues.has(id));
+    expect(missing).toEqual([]);
   });
 
   test('folds counters into Modules delegation without an enable switch', () => {
